Fix for RGB color mismatch for both RGB888 and RGB565 formats
The CDC200 expects the colors to be in little endian formats in Blue,
Green and Red order. For ARGB8888 the copying of color to color buffer
in application which updates the CDC200 framebuffer was working fine,
but for RGB888 and RGB565 the bytes were getting swapped out.
Hence, for RGB888:
Red color component was being displayed as
Blue, while Blue component became Red.
And for RGB565:
Blue was being observed as green,
Green color component was contained a mix of Red and Blue but
majorly red, and
Red component will contain components of Green and Blue, but majorly
Blue.

Signed-off-by: Yogender Kumar Arya <yogender.kumar@alifsemi.com>