samples: drivers: spi_dw: dma support
- Supported the below modes
    - Master Transmit
    - Master Receive
    - Slave Transmit
    - Slave Receive

The master thread will send predefined Tx Data to slave(SPI0)
and similarly Slave will send Tx Data to Master. The result
will be displayed in the console.

Once we have the eventrouter driver, channel configurations
will be moved out of this application.

For Demo, DMA-0 used for E3 and DMA-2 used for E7.
DMA enabled as default.

Signed-off-by: Sudhir Sreedharan <sudhir@alifsemi.com>
Signed-off-by: Ganesh Kumar R <ganesh.ramani@alifsemi.com>