samples: video: Fix cam app for non-HM0360 sensor.
The camera app is failing for the case we use camera other than HM0360.
The issue is because of code using IS_ENABLED() guard code uses some
APIs that are there only in the case of HM0360 sensor. The compiler
during the pre-processing stage is marking the code inside if(0), but
the code is compiled out in later stages of compilation only. So, during
the rest of compilation process, the un-available APIs are tried to be
resolved, which causes a compile time failure

Signed-off-by: Yogender Kumar Arya <yogender.kumar@alifsemi.com>