Add alif_img_class and alif_object_detection samples under modules/tflite-micro
Only alif_object_detection has been actively worked with. It has the following issues: colors are not correct, the camera image is garbled badly, and there's a memory leak or something that hangs the example after a while.

Besides that, there is a need to carry out lot of cleanup and code refactoring.